Corrigiendo migrates

+++ sigo/db/migrate/20140611211641_dropweekly_table.rb
@@ -1,5 +1,5 @@
 class DropweeklyTable < ActiveRecord::Migration
   def change
-  	drop_table :weekly_tables
+  	drop_table :weekly_tables if table_exists? :weekly_tables
   end
 end
